+++ NativeCollections/NativeCollections/NativeConcurrentChunkedStream.cs
@@ -359,7 +359,7 @@ namespace NativeCollections
                                 }
 
                                 --Chunks;
-                                Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *Head->Array, (uint)size);
+                                Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *Head->Array, (uint)size);
                                 byteCount += size;
                             }
 
@@ -379,7 +379,7 @@ namespace NativeCollections
                                 }
 
                                 --Chunks;
-                                Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *Head->Array, (uint)remaining);
+                                Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *Head->Array, (uint)remaining);
                             }
                         }
                     }
@@ -426,7 +426,7 @@ namespace NativeCollections
                             }
 
                             --Chunks;
-                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *Head->Array, (uint)size);
+                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *Head->Array, (uint)size);
                             byteCount += size;
                         }
 
@@ -446,7 +446,7 @@ namespace NativeCollections
                             }
 
                             --Chunks;
-                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, byteCount), ref *Head->Array, (uint)remaining);
+                            Unsafe.CopyBlockUnaligned(ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), ref *Head->Array, (uint)remaining);
                         }
 
                         ReadOffset = remaining;
@@ -505,7 +505,7 @@ namespace NativeCollections
                         Tail->Next = chunk;
                         Tail = chunk;
                         ++Chunks;
-                        Unsafe.CopyBlockUnaligned(ref *Tail->Array, ref Unsafe.AddByteOffset(ref reference, byteCount), (uint)size);
+                        Unsafe.CopyBlockUnaligned(ref *Tail->Array, ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), (uint)size);
                         byteCount += size;
                     }
 
@@ -525,7 +525,7 @@ namespace NativeCollections
                         Tail->Next = chunk;
                         Tail = chunk;
                         ++Chunks;
-                        Unsafe.CopyBlockUnaligned(ref *Tail->Array, ref Unsafe.AddByteOffset(ref reference, byteCount), (uint)remaining);
+                        Unsafe.CopyBlockUnaligned(ref *Tail->Array, ref Unsafe.AddByteOffset(ref reference, (nint)byteCount), (uint)remaining);
                     }
 
                     WriteOffset = remaining;
